move parse-booleanish -> types/src/booleanish, widen args and add tests for null/undefined
