+++ packages/types/index.ts
@@ -1,4 +1,7 @@
+export { parseBooleanish } from "./src/booleanish";
+
 export type { WithAsChild } from "./src/as-child";
 export type { WithStyleProps } from "./src/with-style-props";
 export type { VariantProps } from "./src/variant-props";
 export type { DeepNonNullable } from "./src/deep-non-nullable";
+export type { Booleanish } from "./src/booleanish";
